pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fast-zero"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "André P. Santos", email = "andreztz@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "fastapi>=0.111.0",
    "sqlalchemy>=2.0.31",
    "pydantic-settings>=2.3.4",
    "alembic>=1.13.2",
    "pwdlib[argon2]",
    "python-multipart",
    "pyjwt",
]

[project.urls]
Documentation = "https://github.com/André P. Santos/fast-zero#readme"
Issues = "https://github.com/André P. Santos/fast-zero/issues"
Source = "https://github.com/André P. Santos/fast-zero"

[tool.hatch.version]
path = "fast_zero/__about__.py"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest>=8.2.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.8",
]

[tool.hatch.envs.default.scripts]
server = "fastapi dev fast_zero/app.py"
format = "ruff check . --fix && ruff format ."
lint = "ruff check . && ruff check . --diff"
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
  "coverage html",
]
cov = [
  "test-cov",
  "cov-report",
]


